Xcos: first version using the MVC
No more no less than before : just instantiate a diagram.

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/port/Orientation.java
@@ -50,10 +50,9 @@ public enum Orientation {
      *            The block mirror state
      * @return The value of the angle.
      */
-    public int getRelativeAngle(int blockAngle,
-    Class <? extends BasicPort > klass, boolean flipped, boolean mirrored) {
-        final int orientation = getOrientationAngle();
-        final int base = getBaseAngle(klass, orientation);
+    public double getRelativeAngle(double blockAngle, Class <? extends BasicPort > klass, boolean flipped, boolean mirrored) {
+        final double orientation = getOrientationAngle();
+        final double base = getBaseAngle(klass, orientation);
 
         return getFlippedAndMirroredAngle(base + blockAngle, flipped, mirrored);
     }
@@ -67,10 +66,10 @@ public enum Orientation {
      *            the mirror status
      * @return the real angle
      */
-    public int getAbsoluteAngle(Class <? extends BasicPort > klass,
+    public double getAbsoluteAngle(Class <? extends BasicPort > klass,
     boolean flipped, boolean mirrored) {
-        final int orientation = getOrientationAngle();
-        final int base = getBaseAngle(klass, orientation);
+        final double orientation = getOrientationAngle();
+        final double base = getBaseAngle(klass, orientation);
 
         return getFlippedAndMirroredAngle(base, flipped, mirrored);
     }
@@ -86,9 +85,8 @@ public enum Orientation {
      *            the mirror status
      * @return the updated angle.
      */
-    private int getFlippedAndMirroredAngle(int base, boolean flipped,
-    boolean mirrored) {
-        int angle = base;
+    private double getFlippedAndMirroredAngle(double base, boolean flipped, boolean mirrored) {
+        double angle = base;
 
         switch (this) {
             case NORTH:
@@ -130,8 +128,7 @@ public enum Orientation {
      *            calculated orientation angle
      * @return updated angle
      */
-    private int getBaseAngle(Class <? extends BasicPort > klass,
-                             int orientationAngle) {
+    private double getBaseAngle(Class <? extends BasicPort > klass, double orientationAngle) {
         final boolean isOutput = OutputPort.class.isAssignableFrom(klass)
                                  || CommandPort.class.isAssignableFrom(klass);
